backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
google-cloud-translate
googletrans
pydantic
pydantic-settings
openai
python-multipart
openai-whisper